+++ src/pages/t/profile/change-password.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import ChangePasswordPage from "@/components/profile/ChangePassword";
+
+const changePassword = () => {
+	return <ChangePasswordPage />;
+};
+
+export default changePassword;
