+++ src/pages/t/profile/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Profile from "@/components/profile/Profile";
+
+const ProfilePage = () => {
+	return <Profile />;
+};
+
+export default ProfilePage;
